requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2017.7.27.1
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.42.0
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-extensions==3.1.3
django-js-asset==1.2.2
django-materialize-form==0.1.1
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
graphviz==0.16
hiredis==2.0.0
hyperlink==21.0.0
idna==2.6
incremental==21.3.0
msgpack==1.0.2
oauthlib==3.1.0
olefile==0.44
phonenumbers==8.12.13
Pillow==8.0.1
psycopg2==2.8.6
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
Pygments==2.8.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2017.2
redis==3.5.3
requests==2.18.4
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
twilio==6.52.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.22
validate-email==1.3
zope.interface==5.3.0
